Added code for data and loader classes.
About data.js: Everything is somewhat scaled to Earth's properties.
After Mars the scaling for planet and orbit radiuses is very inaccurate
though but should be fine (otherwise we'd need a 4K+ monitor to fit the
solar system on one screen). After all images are added I'll fix the
sources for the objects in Data. Right now all that are missing have the
Sun's image as a placeholder.
Use Data.collection to access the data.

About loader.js: I'll finish it when I know what the constructors for
Orbit and Planet (or CelestialObject, or whatever) would look like.
Define new Loader(..., ..., ...) and use loader.load() to get the data
in Data.collection as an array of Planet objects (or CelestialBody() or
whatever) with their respective orbits.

Added the Excel file which I used for the scaling of the objects'
properties.