Added info to data.js, added some code to loader.js
The information about celestial objects is in the object 'data'. In
data.info there is a short summary about the planet form Wikipedia. The
other properties are meant to make it easier to make something like
.toString() for the planets if we ever need it.
Added construction of the orbits and space objects in the loader.js
(with a suggestion constructor for the space objects).
Comments in bulgarian are not meant to stay in the code.