Fixed angular and orbital speed in data.js.
I had calculated the angular and orbital speeds wrong. Fixed them.
Fixed the unreadable comments in loader.js... Git clearly does not like
Cyrillic.
In the "data" folder -> Fixed the incorrect data in the excel file.
dataTest.js is almost the same as main.js but with the data from
data.js. This is how the animation would look like with the properties
in data.js.
Also with the new values for angular speed there is no need to divide by
1000 when calculating angleDiff at the end of main/dataTest (rows 425,
430 in main.js or 430, 435 in dataTest.js).